+++ generators/entity/logic/asserts.mjs
