+++ generators/entity/prompts.mjs
@@ -21,7 +21,7 @@ import fs from 'fs';
 import _ from 'lodash';
 import constants from '../generator-constants.cjs';
 import { reservedKeywords, databaseTypes, applicationTypes, entityOptions, fieldTypes, validations } from '../../jdl/jhipster/index.mjs';
-import { inputIsNumber, inputIsSignedDecimalNumber, inputIsSignedNumber } from '../base/logic/index.mjs';
+import { inputIsNumber, inputIsSignedDecimalNumber, inputIsSignedNumber } from './logic/index.mjs';
 
 const { isReservedPaginationWords, isReservedFieldName, isReservedTableName } = reservedKeywords;
 const { CASSANDRA, SQL } = databaseTypes;
